api/requirements.txt

flask
gunicorn
psycopg2-binary
pika
Flask-SQLAlchemy